moved c code to external sources
